Add dataset comments

+++ backend/fms_core/models/property_value.py
@@ -22,7 +22,7 @@ class PropertyValue(TrackedModel):
     property_type = models.ForeignKey(PropertyType, on_delete=models.PROTECT, related_name="property_values", help_text="Property type.")
     content_type_choices = models.Q(app_label='fms_core', model='process') | models.Q(app_label='fms_core', model='processmeasurement')
     content_type = models.ForeignKey(ContentType, on_delete=models.PROTECT, limit_choices_to=content_type_choices)
-    object_id = models.PositiveIntegerField()
+    object_id = models.BigIntegerField()
     content_object = GenericForeignKey('content_type', 'object_id')
 
     def clean(self):
